Remember to update prefabs !
The grapple cooldown was not working properly in other scenes because I had not applied the changes to the prefab in my scene. It should be working properly now. Also added a temp sprite for testing the character flip function and also made the UI element for the grapple stacks just a bit better.(not ready yet)